refactor: list-dependencies 処理を別ファイル化
他の機能からも呼び出す準備
